Added gitignore as pyc files were showing up.
gitconfig
Adds difftool and mergetool as sourcetree.
Forces ssh instead of https.
Changes default branch to 'main'.

profile
Made the pyenv lines conditional in pyenv is installed.
Added ruby PATH and FLAG changes.
Added a newline to the prompt.
Added UID/GID to fix an oddity in postgresql/docker.  Only applied when sharing repo with linux and having the data dir stored as local files.

vimrc
I believe this is to set width of spaces for sh files.  I don't recall why I set this.  It was 1.5yrs ago.